fix: move types, add delete, fix styles
