+++ go.mod
@@ -6,7 +6,7 @@ toolchain go1.21.4
 
 require (
 	github.com/jpillora/overseer v1.1.6
-	github.com/mishamyrt/nuga-lib v0.7.2
+	github.com/mishamyrt/nuga-lib v0.7.3
 	github.com/wailsapp/wails/v2 v2.8.0
 )
 
@@ -39,7 +39,7 @@ require (
 	golang.org/x/crypto v0.18.0 // indirect
 	golang.org/x/exp v0.0.0-20230522175609-2e198f4a06a1 // indirect
 	golang.org/x/net v0.20.0 // indirect
-	golang.org/x/sys v0.18.0 // indirect
+	golang.org/x/sys v0.19.0 // indirect
 	golang.org/x/text v0.14.0 // indirect
 )
