+++ ui/src/entities/keys/lib/index.ts
@@ -1,5 +1,6 @@
 export * from './action'
 export * from './build-template'
 export * from './constants'
+export * from './events'
 export * from './key-name'
 export * from './macros'
