+++ ui/src/entities/keys/model/types/macro.ts
@@ -15,33 +15,6 @@ export type Macro = {
   title: string
 }
 
-export const enum MacroStepType {
-  Wait = 'Wait',
-  KeyDown = 'Key down',
-  KeyUp = 'Key up',
-}
-
-export type MacroKeyStepType = MacroStepType.KeyDown | MacroStepType.KeyUp
-
-export type MacroBasicStep = {
-  type: MacroStepType
-  id: string
-  keyName?: string
-  delay?: number
-}
-
-export type MacroWaitStep = {
-  type: MacroStepType.Wait
-  delay: number
-} & MacroBasicStep
-
-export type MacroKeyStep = {
-  type: MacroKeyStepType
-  keyName: string
-} & MacroBasicStep
-
-export type MacroStep = MacroWaitStep | MacroKeyStep
-
 export type MacroChangedParams = {
   index: number
   title: string
