+++ ui/src/features/keys/record-keystroke/lib/index.ts
@@ -1,2 +1,1 @@
 export * from './constants'
-export * from './event'
