+++ ui/src/features/keys/select-macro/lib/index.ts
@@ -1,1 +1,2 @@
-export * from './step-observer'
+export * from './observer'
+export * from './steps'
