+++ ui/src/features/keys/select-macro/model/index.ts
@@ -1,1 +1,2 @@
 export * from './store'
+export * from './types'
