+++ ui/src/features/keys/select-macro/model/types.ts
@@ -1,4 +1,29 @@
-import type { MacroStep } from '$entities/keys'
+export const enum MacroStepType {
+  Wait = 'Wait',
+  KeyDown = 'Key down',
+  KeyUp = 'Key up',
+}
+
+export type MacroKeyStepType = MacroStepType.KeyDown | MacroStepType.KeyUp
+
+export type MacroBasicStep = {
+  type: MacroStepType
+  id: string
+  keyName?: string
+  delay?: number
+}
+
+export type MacroWaitStep = {
+  type: MacroStepType.Wait
+  delay: number
+} & MacroBasicStep
+
+export type MacroKeyStep = {
+  type: MacroKeyStepType
+  keyName: string
+} & MacroBasicStep
+
+export type MacroStep = MacroWaitStep | MacroKeyStep
 
 export type CustomDragEvent = CustomEvent<{
   items: MacroStep[]
